refactor: VerificationEmailController 의존 관계 리팩터링
