+++ src/test/java/com/jobnote/domain/user/service/UserServiceTest.java
@@ -6,7 +6,6 @@ import com.jobnote.domain.email.domain.VerificationEmailFixture;
 import com.jobnote.domain.email.domain.VerificationEmailType;
 import com.jobnote.domain.user.domain.User;
 import com.jobnote.domain.user.domain.UserFixture;
-import com.jobnote.domain.user.domain.UserRole;
 import com.jobnote.domain.user.dto.*;
 import com.jobnote.domain.email.domain.VerificationEmail;
 import com.jobnote.domain.user.repository.UserRepository;
@@ -22,7 +21,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
-import java.util.UUID;
 
 import static com.jobnote.global.common.ResponseCode.*;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -217,28 +215,6 @@ class UserServiceTest extends ServiceUnitTest {
             assertThatThrownBy(() -> userService.getUserById(userId))
                     .isInstanceOf(JobNoteException.class)
                     .hasMessage(ResponseCode.NOT_FOUND_USER.getMessage());
-        }
-    }
-
-    @Nested
-    @DisplayName("이메일 인증")
-    class EmailVerification {
-        @Test
-        @DisplayName("성공 - 회원의 Role은 MEMBER가 된다.")
-        void success() {
-            // given
-            final User user = UserFixture.createGuest("testEmail@test.com", "testPassword", "testNickname");
-            final LocalDateTime expiryDate = LocalDateTime.of(2025, 7, 30, 12, 0);
-            final String token = UUID.randomUUID().toString();
-            final VerificationEmail verificationEmail = VerificationEmailFixture.createVerifiedSignUp(token, user, expiryDate);
-
-            given(verificationEmailService.verify(token)).willReturn(verificationEmail);
-
-            // when
-            userService.verifySignUp(token);
-
-            // then
-            assertThat(user.getRole()).isEqualTo(UserRole.MEMBER);
         }
     }
